requirements.txt
openai
python-dotenv
google-generativeai
googlesearch-python
